requirements.txt
tensorflow
tensorflow-text
streamlit
pytesseract
Pillow
PyMuPDF
<<<<<<< HEAD
=======
tesseract-ocr
>>>>>>> dca7a2f (handle tesseract exceptions)
